requirements.txt
pandas~=1.1.0
matplotlib~=3.2.2
sklearn~=0.0
scikit-learn~=0.23.2
tqdm~=4.48.2
scipy~=1.4.1
lightgbm~=2.3.1
joblib~=0.16.0
tensorflow-gpu~=2.3.0
statsmodels==0.11.1
tsfresh~=0.16.0
imbalanced-learn~=0.6.2